updated to use Winston and LogDNA-Winston logger.  Handled empty Chrono return
Changed from @logdna/logger to use winston and logdna-winston libraries.  Dev logs are piped and formatted to the console by default and if LOGDNA_API_KEY env var is set, they will also be piped to LogDNA via logdna-winston.

Handled case where chrono.parse returns an empty array (ie, bad time specification).  Ignoring the specified option when this happens and outputting a log regardless if it's dev or not.