feat(tail): Include websocket pings to keep socket up behind proxies
When the remote livetail service is behind an aggresive proxy, it is
likely that idle connections will be terminated. While normal behavior,
it can make for a less than ideal experience you have the connection
recycling every minute.

This adds a ping interval to send ping packets as described in the
websocket spec to keep the socket connection from being disconnected
while idle

Ref: LOG-11362